Development

+++ avail-server/src/main/kotlin/com/avail/server/messages/binary/editor/FileOpenedMessage.kt
@@ -6,16 +6,16 @@
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
  *
- *  * Redistributions of source code must retain the above copyright notice, this
- *     list of conditions and the following disclaimer.
+ * * Redistributions of source code must retain the above copyright notice, this
+ *   list of conditions and the following disclaimer.
  *
- *  * Redistributions in binary form must reproduce the above copyright notice, this
- *     list of conditions and the following disclaimer in the documentation
- *     and/or other materials provided with the distribution.
+ * * Redistributions in binary form must reproduce the above copyright notice,
+ *   this list of conditions and the following disclaimer in the documentation
+ *   and/or other materials provided with the distribution.
  *
- *  * Neither the name of the copyright holder nor the names of the contributors
- *    may be used to endorse or promote products derived from this software
- *    without specific prior written permission.
+ * * Neither the name of the copyright holder nor the names of the contributors
+ *   may be used to endorse or promote products derived from this software
+ *   without specific prior written permission.
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
@@ -33,12 +33,10 @@
 package com.avail.server.messages.binary.editor
 
 import com.avail.server.io.AvailServerChannel
-import com.avail.server.io.files.FileManager
+import com.avail.files.FileManager
 import com.avail.server.messages.Message
 import com.avail.server.session.Session
-import java.nio.ByteBuffer
 import java.nio.charset.StandardCharsets
-import java.util.*
 
 /**
  * `FileOpenMessage` is a [BinaryMessage] response to an open file request.
@@ -64,7 +62,7 @@ import java.util.*
 internal class FileOpenedMessage constructor(
 	override var commandId: Long,
 	fileId: Int,
-	fileSize: Int,
+	fileSize: Long,
 	mime: String): BinaryMessage()
 {
 	override val command = BinaryCommand.FILE_OPENED
@@ -82,7 +80,7 @@ internal class FileOpenedMessage constructor(
 		val bufferSize = 24 + mimeBytes.size
 		val buffer = buffer(bufferSize)
 		buffer.putInt(fileId)
-		buffer.putInt(fileSize)
+		buffer.putLong(fileSize)
 		buffer.putInt(mimeBytes.size)
 		buffer.put(mimeBytes)
 		buffer.flip()
